create capaian luaran for role dosen and role academic

+++ routes/web.php
@@ -8,6 +8,8 @@ use App\Http\Controllers\PenelitianController;
 use App\Http\Controllers\PengabdianController;
 use App\Http\Controllers\PengajaranController;
 use App\Http\Controllers\PesanController;
+use App\Http\Controllers\MenuPenunjangController;
+use App\Http\Controllers\CapaianLuaranController;
 use App\Http\Controllers\ProfileController;
 use App\Http\Controllers\PusatInformasiController;
 use App\Models\Pengabdian;
@@ -30,7 +32,36 @@ use Spatie\Permission\Models\Role;
 */
 
 Route::get('/', function () {
-    return view('welcome');
+    return view('home', [
+        "title" => "Beranda"
+    ]);
+});
+
+Route::get('/about', function () {
+    return view('about', [
+        "title" => "Tentang"
+    ]);
+});
+
+Route::get('/dosen', function () {
+    return view('dosen', [
+        "title" => "Dosen"
+    ]);
+});
+
+Route::get('/detaildosen', function () {
+    return view('detaildosen', [
+        "title" => "Detail Dosen"
+    ]);
+});
+
+Route::get('/informationcenter', [PusatInformasiController::class, 'show']);
+
+Route::get('/detail_informationcenter', function () {
+    // return view('detail_informationcenter');
+    return view('detail_informationcenter', [
+        "title" => "Pusat Informasi"
+    ]);
 });
 
 Auth::routes([
@@ -40,10 +71,10 @@ Auth::routes([
 // Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
 
 Route::group(['middleware' => ['web', 'auth']], function () {
-    
+
     // Home
-    Route::get('/home', [HomeController::class, 'index'])->middleware('role:dosen|akademik')->name('home');
-    
+    Route::get('/dashboard/home', [HomeController::class, 'index'])->middleware('role:dosen|akademik')->name('home');
+
     // Data Diri
     Route::get('/data_diri', [DataDiriController::class, 'index'])->middleware('role:dosen')->name('data_diri');
 
@@ -51,7 +82,7 @@ Route::group(['middleware' => ['web', 'auth']], function () {
     Route::get('/data_diri/alamat_kontak/{id}/edit', [DataDiriController::class, 'edit_alamat_kontak'])->middleware('role:dosen')->name('alamat_kontak.edit');
     Route::get('/data_diri/kepegawaian/{id}/edit', [DataDiriController::class, 'edit_kepegawaian'])->middleware('role:dosen')->name('kepegawaian.edit');
     Route::get('/data_diri/lain/{id}/edit', [DataDiriController::class, 'edit_lain'])->middleware('role:dosen')->name('lain.edit');
-    
+
     Route::put('/data_diri/identitas/update/{id}', [DataDiriController::class, 'update_identitas'])->middleware('role:dosen')->name('identitas.update');
     Route::put('/data_diri/lain/update/{id}', [DataDiriController::class, 'update_lain'])->middleware('role:dosen')->name('lain.update');
     Route::put('/data_diri/alamat_kontak/update/{id}', [DataDiriController::class, 'update_alamat_kontak'])->middleware('role:dosen')->name('alamat_kontak.update');
@@ -65,13 +96,13 @@ Route::group(['middleware' => ['web', 'auth']], function () {
     Route::get('/pengajaran/{id}/view', [PengajaranController::class, 'view'])->middleware('role:dosen')->name('pengajaran.view');
     Route::put('/pengajaran/update/{id}', [PengajaranController::class, 'update'])->middleware('role:dosen')->name('pengajaran.update');
     Route::delete('/pengajaran/delete/{id}', [PengajaranController::class, 'destroy'])->middleware('role:dosen')->name('pengajaran.destroy');
-    
-    
+
+
     // Profile
     Route::get('/profile', [ProfileController::class, 'index'])->middleware('role:dosen|akademik')->name('profile');
     Route::put('/profile/{id}', [ProfileController::class, 'update'])->middleware('role:dosen|akademik')->name('profile.update');
     // Route::resources('/profile', ProfileController::class);
-    
+
     // Penelitian
     Route::get('/penelitian', [PenelitianController::class, 'index'])->middleware('role:dosen')->name('penelitian');
     Route::get('/penelitian/create', [PenelitianController::class, 'create'])->middleware('role:dosen')->name('penelitian.create');
@@ -80,8 +111,8 @@ Route::group(['middleware' => ['web', 'auth']], function () {
     Route::get('/penelitian/{id}/view', [PenelitianController::class, 'view'])->middleware('role:dosen')->name('penelitian.view');
     Route::put('/penelitian/update/{id}', [PenelitianController::class, 'update'])->middleware('role:dosen')->name('penelitian.update');
     Route::delete('/penelitian/delete/{id}', [PenelitianController::class, 'destroy'])->middleware('role:dosen')->name('penelitian.destroy');
-    
-    
+
+
     // Pendidikan
     Route::get('/pendidikan', [PendidikanController::class, 'index'])->middleware('role:dosen')->name('pendidikan');
     Route::get('/pendidikan/create', [PendidikanController::class, 'create'])->middleware('role:dosen')->name('pendidikan.create');
@@ -90,7 +121,7 @@ Route::group(['middleware' => ['web', 'auth']], function () {
     Route::get('/pendidikan/{id}/view', [Pendidikan::class, 'view'])->middleware('role:dosen')->name('pendidikan.view');
     Route::put('/pendidikan/update/{id}', [PendidikanController::class, 'update'])->middleware('role:dosen')->name('pendidikan.update');
     Route::delete('/pendidikan/delete/{id}', [PendidikanController::class, 'destroy'])->middleware('role:dosen')->name('pendidikan.destroy');
-    
+
     // Pengabdian
     Route::get('/pengabdian', [PengabdianController::class, 'index'])->middleware('role:dosen')->name('pengabdian');
     Route::get('/pengabdian/create', [PengabdianController::class, 'create'])->middleware('role:dosen')->name('pengabdian.create');
@@ -99,40 +130,71 @@ Route::group(['middleware' => ['web', 'auth']], function () {
     Route::get('/pengabdian/{id}/view', [PengabdianController::class, 'view'])->middleware('role:dosen')->name('pengabdian.view');
     Route::put('/pengabdian/update/{id}', [PengabdianController::class, 'update'])->middleware('role:dosen')->name('pengabdian.update');
     Route::delete('/pengabdian/delete/{id}', [PengabdianController::class, 'destroy'])->middleware('role:dosen')->name('pengabdian.destroy');
-    
+
+    // Menu Penunjang
+    Route::get('/menu_penunjang', [MenuPenunjangController::class, 'index'])->middleware('role:dosen')->name('menu_penunjang');
+    Route::get('/menu_penunjang/create', [MenuPenunjangController::class, 'create'])->middleware('role:dosen')->name('menu_penunjang.create');
+    Route::post('/menu_penunjang/create', [MenuPenunjangController::class, 'store'])->middleware('role:dosen')->name('menu_penunjang.store');
+    Route::get('/menu_penunjang/{id}/edit', [MenuPenunjangController::class, 'edit'])->middleware('role:dosen')->name('menu_penunjang.edit');
+    Route::put('/menu_penunjang/update/{id}', [MenuPenunjangController::class, 'update'])->middleware('role:dosen')->name('menu_penunjang.update');
+    Route::delete('/menu_penunjang/delete/{id}', [MenuPenunjangController::class, 'destroy'])->middleware('role:dosen')->name('menu_penunjang.destroy');
+
+    // Capaian Luaran
+    Route::get('/capaian_luaran', [CapaianLuaranController::class, 'index'])->middleware('role:dosen')->name('capaian_luaran');
+    Route::get('/capaian_luaran/create', [CapaianLuaranController::class, 'create'])->middleware('role:dosen')->name('capaian_luaran.create');
+    Route::post('/capaian_luaran/create', [CapaianLuaranController::class, 'store'])->middleware('role:dosen')->name('capaian_luaran.store');
+    Route::get('/capaian_luaran/{id}/edit', [CapaianLuaranController::class, 'edit'])->middleware('role:dosen')->name('capaian_luaran.edit');
+    Route::put('/capaian_luaran/update/{id}', [CapaianLuaranController::class, 'update'])->middleware('role:dosen')->name('capaian_luaran.update');
+    Route::delete('/capaian_luaran/delete/{id}', [CapaianLuaranController::class, 'destroy'])->middleware('role:dosen')->name('capaian_luaran.destroy');
+
     // ROLE AKADEMIK
-    Route::get('/dosen', [DosenController::class, 'index'])->middleware('role:akademik')->name('dosen');
+    Route::get('/dosen/all', [DosenController::class, 'index'])->middleware('role:akademik')->name('dosen');
     Route::get('/dosen/search', [DosenController::class, 'search'])->middleware('role:akademik')->name('dosen.search');
     Route::get('/dosen/{id}/edit', [DosenController::class, 'edit'])->middleware('role:akademik')->name('dosen.edit');
-    
+
     // Pendidikan
     Route::get('/dosen/{id}/pendidikan/create', [DosenController::class, 'create_pendidikan'])->middleware('role:akademik')->name('dosen.pendidikan.create');
     Route::post('/dosen/{id}/pendidikan/create', [DosenController::class, 'store_pendidikan'])->middleware('role:akademik')->name('dosen.pendidikan.store');
     Route::get('/dosen/{user_id}/pendidikan/{id}/edit', [DosenController::class, 'edit_pendidikan'])->middleware('role:akademik')->name('dosen.pendidikan.edit');
     Route::put('/dosen/{user_id}/pendidikan/update/{id}', [DosenController::class, 'update_pendidikan'])->middleware('role:akademik')->name('dosen.pendidikan.update');
     Route::delete('/dosen/pendidikan/delete/{id}', [DosenController::class, 'destroy_pendidikan'])->middleware('role:akademik')->name('dosen.pendidikan.destroy');
-    
+
     // Pengajaran
     Route::get('/dosen/{id}/pengajaran/create', [DosenController::class, 'create_pengajaran'])->middleware('role:akademik')->name('dosen.pengajaran.create');
     Route::post('/dosen/{id}/pengajaran/create', [DosenController::class, 'store_pengajaran'])->middleware('role:akademik')->name('dosen.pengajaran.store');
     Route::get('/dosen/{user_id}/pengajaran/{id}/edit', [DosenController::class, 'edit_pengajaran'])->middleware('role:akademik')->name('dosen.pengajaran.edit');
     Route::put('/dosen/{user_id}/pengajaran/update/{id}', [DosenController::class, 'update_pengajaran'])->middleware('role:akademik')->name('dosen.pengajaran.update');
-    Route::delete('/dosen/pengajaran/delete/{id}', [DosenController::class, 'destroy_pengajaran'])->middleware('r ole:akademik')->name('dosen.pengajaran.destroy');
-    
+    Route::delete('/dosen/{user_id}/pengajaran/delete/{id}', [DosenController::class, 'destroy_pengajaran'])->middleware('r ole:akademik')->name('dosen.pengajaran.destroy');
+
     // Penelitian
     Route::get('/dosen/{id}/penelitian/create', [DosenController::class, 'create_penelitian'])->middleware('role:akademik')->name('dosen.penelitian.create');
     Route::post('/dosen/{id}/penelitian/create', [DosenController::class, 'store_penelitian'])->middleware('role:akademik')->name('dosen.penelitian.store');
     Route::get('/dosen/{user_id}/penelitian/{id}/edit', [DosenController::class, 'edit_penelitian'])->middleware('role:akademik')->name('dosen.penelitian.edit');
     Route::put('/dosen/{user_id}/penelitian/update/{id}', [DosenController::class, 'update_penelitian'])->middleware('role:akademik')->name('dosen.penelitian.update');
-    Route::delete('/dosen/penelitian/delete/{id}', [DosenController::class, 'destroy_penelitian'])->middleware('role:akademik')->name('dosen.penelitian.destroy');
-    
+    Route::delete('/dosen/{user_id}/penelitian/delete/{id}', [DosenController::class, 'destroy_penelitian'])->middleware('role:akademik')->name('dosen.penelitian.destroy');
+
     // Pengabdian
     Route::get('/dosen/{id}/pengabdian/create', [DosenController::class, 'create_pengabdian'])->middleware('role:akademik')->name('dosen.pengabdian.create');
     Route::post('/dosen/{id}/pengabdian/create', [DosenController::class, 'store_pengabdian'])->middleware('role:akademik')->name('dosen.pengabdian.store');
     Route::get('/dosen/{user_id}/pengabdian/{id}/edit', [DosenController::class, 'edit_pengabdian'])->middleware('role:akademik')->name('dosen.pengabdian.edit');
     Route::put('/dosen/{user_id}/pengabdian/update/{id}', [DosenController::class, 'update_pengabdian'])->middleware('role:akademik')->name('dosen.pengabdian.update');
-    Route::delete('/dosen/pengabdian/delete/{id}', [DosenController::class, 'destroy_pengabdian'])->middleware('role:akademik')->name('dosen.pengabdian.destroy');
-    
+    Route::delete('/dosen/{user_id}/pengabdian/delete/{id}', [DosenController::class, 'destroy_pengabdian'])->middleware('role:akademik')->name('dosen.pengabdian.destroy');
+
+    // Penunjang
+    Route::get('/dosen/{id}/penunjang/create', [DosenController::class, 'create_penunjang'])->middleware('role:akademik')->name('dosen.penunjang.create');
+    Route::post('/dosen/{id}/penunjang/create', [DosenController::class, 'store_penunjang'])->middleware('role:akademik')->name('dosen.penunjang.store');
+    Route::get('/dosen/{user_id}/penunjang/{id}/edit', [DosenController::class, 'edit_penunjang'])->middleware('role:akademik')->name('dosen.penunjang.edit');
+    Route::put('/dosen/{user_id}/penunjang/update/{id}', [DosenController::class, 'update_penunjang'])->middleware('role:akademik')->name('dosen.penunjang.update');
+    Route::delete('/dosen/{user_id}/penunjang/delete/{id}', [DosenController::class, 'destroy_penunjang'])->middleware('role:akademik')->name('dosen.penunjang.destroy');
+
+    // Capaian Luaran
+    Route::get('/dosen/{id}/capaian_luaran/create', [DosenController::class, 'create_capaian_luaran'])->middleware('role:akademik')->name('dosen.capaian_luaran.create');
+    Route::post('/dosen/{id}/capaian_luaran/create', [DosenController::class, 'store_capaian_luaran'])->middleware('role:akademik')->name('dosen.capaian_luaran.store');
+    Route::get('/dosen/{user_id}/capaian_luaran/{id}/edit', [DosenController::class, 'edit_capaian_luaran'])->middleware('role:akademik')->name('dosen.capaian_luaran.edit');
+    Route::put('/dosen/{user_id}/capaian_luaran/update/{id}', [DosenController::class, 'update_capaian_luaran'])->middleware('role:akademik')->name('dosen.capaian_luaran.update');
+    Route::delete('/dosen/{user_id}/capaian_luaran/delete/{id}', [DosenController::class, 'destroy_capaian_luaran'])->middleware('role:akademik')->name('dosen.capaian_luaran.destroy');
+
+
     // Data Diri
     Route::get('/dosen/{user_id}/identitas/{id}/edit', [DosenController::class, 'edit_identitas'])->middleware('role:akademik')->name('dosen.identitas.edit');
     Route::get('/dosen/{user_id}/kepegawaian/{id}/edit', [DosenController::class, 'edit_kepegawaian'])->middleware('role:akademik')->name('dosen.kepegawaian.edit');
